+++ frontend/src/App.jsx
@@ -1,15 +1,16 @@
-import Canvas from "./components/Canvas";
-import { useFontLoader } from "./hooks/useFontLoader";
+import NavBar from "./components/NavBar";
+import Home from "./pages/Home";
+import Tool1 from "./pages/Tool1";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 
 export default function App() {
-	const { font, loading } = useFontLoader("/fonts/Helvetica/Helvetica.ttf");
-
-	if (loading) return <p>Font wordt geladen…</p>;
-
 	return (
-		<div style={{ padding: "2rem" }}>
-			{/* Geef Canvas het geladen font en een test-letter */}
-			<Canvas font={font} selectedChar="A" />
-		</div>
+		<Router>
+			<NavBar />
+			<Routes>
+				<Route path="/" element={<Home />} />{" "}
+				<Route path="/tool1" element={<Tool1 />} />
+			</Routes>
+		</Router>
 	);
 }